Get rid of deprecated Gdk.threads_init/enter/leave
These are deprecated and were throwing error messages. In the process,
refactor how the man page index is generated. It used to be that
this was done when the index toolbar button or the respective menu
item were clicked and a window with a pulsing progress bar was
shown. Now it is done when gtkman starts and a statusbar has been
added with text to indicate what's beeing done. The toolbar
button/menuitem are disabled until the index has been built. It's
clearer this way.

Fixes #7.